Modify: 메인 페이지 컴포넌트 분리 - Feed, Comment, Aside, Recommendation 컴포넌트 js파일, scss파일로 분리
