+++ src/pages/kyeom/Main/Main.scss
@@ -52,68 +52,6 @@ img {
   font-size: 14px;
 }
 
-.Feed {
-  margin-right: 28px;
-
-  .feed-wrapper {
-    @include border-gray;
-    max-width: 614px;
-    background-color: white;
-
-    section:not(.feed-img) {
-      padding-right: 16px;
-      padding-left: 16px;
-    }
-
-    .feed-bar {
-      @include flex-justify-between;
-      padding: 14px 16px;
-
-      header {
-        @include flex-justify-between;
-      }
-
-      .profile-img {
-        @include profile-img(36px);
-        margin-right: 14px;
-      }
-    }
-
-    .button {
-      width: 24px;
-      height: 24px;
-    }
-  }
-}
-
-/* 좋아요, 댓글, 공유, 저장된 항목 버튼 */
-.button-wrapper {
-  @include flex-justify-between;
-  padding: 8px 16px;
-
-  .button-left {
-    @include flex-justify-between;
-  }
-
-  .button {
-    margin: 8px;
-
-    &:first-child {
-      margin-left: 0;
-    }
-
-    &:last-child {
-      margin-right: 0;
-    }
-  }
-
-  .fa-heart {
-    padding-top: 2px;
-    font-size: 22px;
-    cursor: pointer;
-  }
-}
-
 .fa-heart {
   &.far {
     color: #5e5e5e;
@@ -121,156 +59,5 @@ img {
 
   &.fas {
     color: rgb(237, 73, 86);
-  }
-}
-
-/* 좋아요 알림 메세지 */
-.message {
-  margin-bottom: 12px;
-
-  span {
-    font-weight: 600;
-  }
-}
-
-/* 피드 내용 */
-section.content {
-  margin-bottom: 12px;
-
-  .name {
-    margin-right: 8px;
-  }
-}
-
-/* 업로드된 댓글 */
-section.uploaded-comment {
-  margin-bottom: 12px;
-
-  .uploaded-comment-wrapper {
-    display: flex;
-    margin-bottom: 8px;
-
-    .name {
-      margin-right: 8px;
-    }
-
-    .uploaded-comment {
-      flex-grow: 1;
-      overflow-wrap: break-word;
-    }
-
-    .delete {
-      all: unset;
-      margin-right: 8px;
-      color: $gray3;
-      font-size: 12px;
-      font-weight: 600;
-      cursor: pointer;
-    }
-
-    .like {
-      font-size: 12px;
-      cursor: pointer;
-
-      i {
-        vertical-align: middle;
-      }
-    }
-  }
-}
-
-/* 댓글 달기 */
-section.comment {
-  padding: 6px 16px;
-  border-top: 1px solid $gray2;
-
-  form {
-    @include flex-justify-between;
-
-    input.comment {
-      flex-grow: 1;
-      height: 36px;
-      margin-right: 8px;
-      border: none;
-    }
-  }
-
-  button {
-    @include button-blue(14px);
-
-    &:disabled {
-      color: $lightblue;
-    }
-  }
-}
-
-/* aside */
-.Aside {
-  position: fixed;
-  top: (32px + $desktop-nav-height);
-  width: 293px;
-
-  .user-wrapper {
-    display: flex;
-    align-items: center;
-    margin: 18px auto 10px;
-
-    .user-img {
-      @include profile-img(64px);
-    }
-
-    .user-profile {
-      flex-grow: 1;
-      margin-left: 12px;
-
-      .name {
-        margin-bottom: 4px;
-
-        & + div {
-          color: $gray3;
-        }
-      }
-    }
-
-    button {
-      @include button-blue(12px);
-    }
-  }
-}
-
-/* 회원님을 위한 추천 */
-.recommendation-wrapper {
-  padding-top: 12px;
-
-  .recommendation-header {
-    @include flex-justify-between;
-    font-weight: 600;
-    color: $gray3;
-
-    button {
-      font-size: 12px;
-    }
-  }
-
-  .recommendation-section {
-    padding: 8px 0 8px 4px;
-
-    .Recommendation {
-      @include flex-justify-between;
-      padding: 8px 0;
-
-      .profile-img {
-        @include profile-img(32px);
-      }
-
-      .name {
-        flex-grow: 1;
-        margin-left: 8px;
-      }
-
-      button {
-        @include button-blue(12px);
-      }
-    }
   }
 }
